% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_pipeline_r.R
\name{factory_pipeline_r}
\alias{factory_pipeline_r}
\title{Pipeline}
\usage{
factory_pipeline_r(
  x,
  y,
  tknz = "spacy",
  ordinal = FALSE,
  metric = "class_balance_accuracy_score",
  cv = 5,
  n_iter = 100,
  n_jobs = 5,
  verbose = 3,
  learners = c("SGDClassifier", "RidgeClassifier", "Perceptron",
    "PassiveAggressiveClassifier", "BernoulliNB", "ComplementNB", "MultinomialNB",
    "RandomForestClassifier"),
  theme = NULL,
  python_setup = FALSE,
  sys_setenv = NULL,
  which_python = NULL,
  which_venv = NULL,
  venv_name = NULL
)
}
\arguments{
\item{venv_name}{}
}
\value{

}
\description{
Pipeline
}
