% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_classification_pipeline_r.R
\name{text_classification_pipeline_r}
\alias{text_classification_pipeline_r}
\title{Whole pipeline}
\usage{
text_classification_pipeline_r(
  filename,
  target,
  predictor,
  test_size = 0.33,
  ordinal = FALSE,
  tknz = "spacy",
  metric = "class_balance_accuracy_score",
  cv = 5,
  n_iter = 100,
  n_jobs = 5,
  verbose = 3,
  learners = c("SGDClassifier"),
  objects_to_save = c("pipeline", "tuning results", "predictions",
    "accuracy per class", "index - training data", "index - test data", "bar plot"),
  save_objects_to_server = FALSE,
  save_objects_to_disk = FALSE,
  save_pipeline_as = "default",
  results_folder_name = "results",
  reduce_criticality = FALSE,
  theme = NULL,
  python_setup = FALSE,
  sys_setenv,
  which_python,
  which_venv,
  venv_name
)
}
\arguments{
\item{venv_name}{}
}
\value{

}
\description{
Whole pipeline
}
