% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_data_load_and_split_r.R
\name{factory_data_load_and_split_r}
\alias{factory_data_load_and_split_r}
\title{Split dataset into training and test sets}
\usage{
factory_data_load_and_split_r(
  filename,
  target,
  predictor,
  test_size = 0.33,
  reduce_criticality = FALSE,
  theme = NULL,
  python_setup = FALSE,
  sys_setenv = NULL,
  which_python = NULL,
  which_venv = NULL,
  venv_name = NULL
)
}
\arguments{
\item{filename}{A string with the dataset name (CSV), including full path to
the data folder (if not in the project's working directory), and the data
type suffix (".csv").}

\item{target}{A string with the name of the response variable.}

\item{predictor}{A string with the name of the predictor variable.}

\item{test_size}{Numeric. Proportion of data that will form the test dataset.}

\item{reduce_criticality}{Logical. For internal use by Nottinghamshire
Healthcare NHS Foundation Trust or other trusts that hold data on
criticality. If \code{TRUE}, then all records with a criticality of "-5"
(respectively, "5") are assigned a criticality of "-4" (respectively, "4").
This is to avoid situations where the pipeline breaks due to a lack of
sufficient data for "-5" and/or "5". Defaults to \code{FALSE}.}

\item{theme}{A string. For internal use by Nottinghamshire Healthcare NHS
Foundation Trust or other trusts that use theme labels ("Access",
"Environment/ facilities" etc.). The column name of the theme variable.
Defaults to \code{NULL}. If supplied, the theme variable will be used as a
predictor (along with the text predictor) in the model that is fitted
with criticality as the response variable. The rationale is two-fold.
First, to help the model improve predictions on criticality when the
theme labels are readily available. Second, to force the criticality for
"Couldn't be improved" to always be "3" in the training and test data, as
well as in the predictions. This is the only criticality value that
"Couldn't be improved" can take, so by forcing it to always be "3", we
are improving model performance, but are also correcting possible
erroneous assignments of values other than "3" that are attributed to
human error.}

\item{python_setup}{A \code{logical} whether to set up the \code{Python} version,
virtual environment etc. that can be controlled with arguments
\code{sys_setenv}, \code{which_python}, \code{which_venv} and \code{venv_name}. These
arguments will be ignored when \code{python_setup} is \code{FALSE}. The purpose of
\code{python_setup} is that users may wish to control the \code{Python} parameters
outside the actual function, for the session in general.}

\item{sys_setenv}{A string in the form "path_to_python/python.exe",
indicating which Python to use (e.g. from a virtual environment).}

\item{which_python}{Same as \code{sys_setenv}.}

\item{which_venv}{A string that can be "conda", "miniconda" or "python".}

\item{venv_name}{String. The name of the virtual environment.}

\item{text_col_name}{A string with the column name of the text variable.}
}
\value{
A list of length 6: \code{x_train} (data frame), \code{x_test} (data frame),
\code{y_train} (character vector), \code{y_test} (character vector), \code{index_train}
(integer vector), and \code{index_test} (integer vector). The row names
(names) of \code{x_train} and \code{x_test} (\code{y_train} and \code{y_test}) are
\code{index_train} and \code{index_test} respectively.
}
\description{
Splits the dataset with \code{Scikit-learn} and returns the train/test data and
their row/position indices.
}
