% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_predict_unlabelled_text_r.R
\name{factory_predict_unlabelled_text_r}
\alias{factory_predict_unlabelled_text_r}
\title{Predict unlabelled text using a fitted \code{Scikit-learn} (Python) pipeline}
\usage{
factory_predict_unlabelled_text_r(
  dataset,
  predictor,
  pipe_path,
  preds_column = NULL,
  column_names = "all_cols",
  theme = NULL
)
}
\arguments{
\item{dataset}{A data frame with the text data to predict classes for.}

\item{predictor}{A string with the column name of the text variable.}

\item{pipe_path}{A string in the form "path_to_fitted_pipeline/pipeline.sav",
where "pipeline" is the name of the SAV file with the fitted
\code{Scikit-learn} pipeline.}

\item{preds_column}{A string with the user-specified name of the column that
will have the predictions. If \code{NULL} (default), then the name will be
\code{paste0(text_col_name, "_preds")}.}

\item{column_names}{A vector of strings with the names of the columns of the
supplied data frame (incl. \code{text_col_name}) to be added to the returned
data frame. If "preds_only", then the only column in the returned data
frame will be \code{preds_column}. Defaults to "all_cols".}

\item{theme}{}
}
\value{

}
\description{
Predict unlabelled text using a fitted \code{Scikit-learn} (Python) pipeline
}
